GetIncoming/Outgoing in Repo returns ChangesetInfo
This commit now updates the GetIncoming & GetOutgoing functions to
return a list of ChangesetInfo namedtuples rather than just a list of
changeset hashes.  This way the user of the Repo class can get access to
the changeset description as well as the hash, to help the user make
decisions about changesets.