Fixed potential issue in Repo._GetChangesetInfoList
Under certain circumstances (misconfigured extensions for example),
mercurial's outgoing/incoming commands will print out an error message
as well as the actual outgoing/incoming changesets.  This was causing
synchg to crash when used, as it expects each line of output to match
the ChangesetInfoRegexp.

This commit fixes that, by filtering out None from the list of
changeset matches.